Prepare schema for push mechanism
In order to implement an incremental update / push mechanism we need to
track database modifications. To that end we add state sequences to all
relevant tables which are maintained by triggers. Currently, only
INSERTs and UPDATEs are considered. DELETE is a bit more involved as we
need to keep tombstones of deleted records.