Completely rework sidebar-resizing.
The new approach relies on one global go-loop handling drag-events sent
over a :shared core.async channe. The incomine messages are then passed
through a multimethod which extracts the target width of the pos-vector
of the drag-element.

This allows us to add resizing to different parts of the application in
a non-complicated and non-intrusive manner.