Implement seqs-handling on the server-side of /poll.
The server will now check if there are any changes for the polling
client. If not, it waits a second and checks again. If there are
changes, the server sends those and the new seq-numbers to the client.

TODO: Wire up the bus so we don't have to poll Postgres.