Remove tags from documents whose files failed in page-extraction.
This is a workaround: If the user uploads a file for which pepa
automatically creates a document, the usual 'default' tags (new,
origin/*, ...) are created. If the page extraction for that document
fails, tags are left on the document.

The query-engine doesn't return documents with 0 pages. That creates an
inconsistent view: The tag list shows one document for tag x, but the
document with that tag doesn't have any pages.